mwlwifi - dont break fw upload timing with log prints
In the function mwl_fwdl_download_firmware() there's a comment that it
takes ~1300 nSec for the register change that acknowledges each chunk
of the firmware uploaded. But if you take too long to get to the check
you can miss the event altogether.

The Function mwl_fwdl_trig_pcicmd_bootcode() contains log prints that
can take several miliseconds to complete and it is called from the above
function to transfer a chunk of the firmware at a time. When logging is
enabled it breaks the firware upload.

So remove the function call that uploads the chunk (and contains log prints)
and add its code in-line to avoid inadvertantly breaking the firmware upload
by enabling logging.

Signed-off-by: Ian Kent <raven@themaw.net>